Rename icall.h to icall-decl.h.
