+++ mono/metadata/Makefile.am
@@ -293,8 +293,8 @@ common_sources = \
 	filewatcher.h		\
 	gc-internals.h		\
 	icall.c			\
-	icall.h			\
 	icalls.h		\
+	icall-decl.h		\
 	icall-internals.h \
 	icall-def.h		\
 	icall-table.h	\
